requirements.txt
uvicorn==0.32.1
fastapi==0.115.5
openai==1.55.0
torch==2.5.1
transformers==4.46.3
langchain==0.3.8
langgraph==0.2.53
qdrant-client==1.12.1
PyPDF2==3.0.1
tiktoken==0.8.0
langchain-text-splitters==0.3.2
langchain_community==0.3.8
langchain-huggingface==0.1.2
sentence_transformers==3.3.1
langchain-qdrant==0.2.0
gradio_client==1.4.3
pydantic==2.10.1
accelerate==1.1.1
streamlit==1.40.1
python-docx==1.1.2
python-pptx==1.0.2
